+++ src/main/java/dev/yanianz/essentials/customscreens/ZCustomScreen.java
@@ -1,4 +1,4 @@
-package fr.maxlego08.essentials.module.modules.customscreens;
+package dev.yanianz.essentials.customscreens;
 
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;